--- project/ovp/frontend/ovp/components/login/index.json
@@ -1,44 +1,0 @@
-{
-  "ko" : {
-    "login" : {
-      "login" : "로그인",
-      "idOrEmail": "아이디 또는 이메일",
-      "validation": {
-        "emailOrId" : {
-          "empty": "사용자 아이디를 입력하세요."
-        },
-        "password" : {
-          "empty": "사용자 비밀번호를 입력하세요."
-        },
-        "fail": "아이디 및 비밀번호가 유효하지 않습니다."
-      },
-      "placeholder": {
-        "idOrEmail": "아이디 또는 이메일 입력",
-        "password": "비밀번호 입력"
-      },
-      "pwFind" : "비밀번호 찾기",
-      "join" : "회원가입"
-    }
-  },
-  "en": {
-    "login" : {
-      "login" : "Login",
-      "idOrEmail": "ID or Email",
-      "validation": {
-        "emailOrId" : {
-          "empty": "Please enter your user ID."
-        },
-        "password" : {
-          "empty": "Please enter your user password."
-        },
-        "fail": "The ID and password are invalid."
-      },
-      "placeholder": {
-        "idOrEmail": "Enter ID or Email",
-        "password": "Enter Password"
-      },
-      "pwFind" : "Find Password",
-      "join" : "Sign Up"
-    }
-  }
-}
